reduceCellRanges
