+++ src/vs/workbench/contrib/notebook/test/notebookCommon.test.ts
@@ -8,7 +8,7 @@ import { Mimes } from 'vs/base/common/mime';
 import { URI } from 'vs/base/common/uri';
 import { IModeService } from 'vs/editor/common/services/modeService';
 import { CellKind, CellUri, diff, NotebookWorkingCopyTypeIdentifier, NOTEBOOK_DISPLAY_ORDER, sortMimeTypes } from 'vs/workbench/contrib/notebook/common/notebookCommon';
-import { cellIndexesToRanges, cellRangesToIndexes } from 'vs/workbench/contrib/notebook/common/notebookRange';
+import { cellIndexesToRanges, cellRangesToIndexes, reduceCellRanges } from 'vs/workbench/contrib/notebook/common/notebookRange';
 import { setupInstantiationService, TestCell } from 'vs/workbench/contrib/notebook/test/testNotebookEditor';
 
 suite('NotebookCommon', () => {
@@ -322,6 +322,12 @@ suite('CellRange', function () {
 
 		assert.deepStrictEqual(cellIndexesToRanges([9, 10]), [{ start: 9, end: 11 }]);
 		assert.deepStrictEqual(cellIndexesToRanges([10, 9]), [{ start: 9, end: 11 }]);
+	});
+
+	test('reduceCellRanges', function () {
+		assert.deepStrictEqual(reduceCellRanges([{ start: 0, end: 2 }, { start: 2, end: 3 }]), [{ start: 0, end: 3 }]);
+		assert.deepStrictEqual(reduceCellRanges([{ start: 2, end: 3 }, { start: 0, end: 2 }]), [{ start: 0, end: 3 }]);
+		assert.deepStrictEqual(reduceCellRanges([{ start: 0, end: 1 }, { start: 2, end: 3 }]), [{ start: 0, end: 1 }, { start: 2, end: 3 }]);
 	});
 });
